Refactor all authoring and data component into a single file
