+++ Assets/Developers/Lab/Components/DayNight.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da8086709d5fac94db1ac741380c0ee9
+guid: 15c1367c3bd2a5e499c5287654888c4d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
