+++ Assets/Developers/Lab/Components/MoveTags.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using Unity.Entities;
+
+namespace JanusVagabond
+{
+  [Serializable]
+  public struct MoveLeft : IComponentData
+  {
+
+  }
+
+  [Serializable]
+  public struct MoveRight : IComponentData
+  {
+
+  }
+}
+
