+++ Assets/Developers/Lab/Components/MoveTags.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8c7a557c8bcf8474a82e7e370702d669
+guid: 3b39e5d9a40f5d64ebcd009c36ebc63d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
