+++ Assets/Developers/Lab/Components/Spawn.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ac7b40f676931b341962ba77c55fc57d
+guid: 4f7b51b38247ab74fa48d50131ebd209
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
