--- Assets/Developers/Lab/Components/SpawnDataComponent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c7779c1fe1ada27468ff8b0bb7b0ae48
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
