+++ Assets/Developers/Lab/Components/Wander.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e95aeefc533421b4aabe9280c18c32b1
+guid: 8e73231ecd99bc84b84f60e638df5f10
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
