+++ Assets/Developers/Lab/Systems/DayNightSystem.cs
@@ -9,7 +9,7 @@ namespace JanusVagabond
 {
 
   [UpdateInGroup(typeof(SimulationSystemGroup))]
-  public class SpawnSystem : JobComponentSystem
+  public class DayNightSystem : JobComponentSystem
   {
     // This declares a new kind of job, which is a unit of work to do.
     // The job is declared as an IJobForEach<Translation, Rotation>,
