+++ Assets/Developers/Lab/Systems/DayNightSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 82ea4d229c729fc46aa9c2ca61275cb6
+guid: 7a7cf6109eff4a74a9a787147ea16728
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
